static_cast t unsigned in addMSecs

+++ src/Pt/Time.cpp
@@ -118,12 +118,12 @@ Time Time::addMSecs(Pt::int64_t ms) const
     Time t;
     if (ms < 0)
     {
-        int negdays = (MSECS_PER_DAY - ms) / MSECS_PER_DAY;
-        t._msecs = (_msecs + ms + negdays * MSECS_PER_DAY) % MSECS_PER_DAY;
+        int negdays = static_cast<int>( (MSECS_PER_DAY - ms) / MSECS_PER_DAY );
+        t._msecs = static_cast<unsigned>( (_msecs + ms + negdays * MSECS_PER_DAY) % MSECS_PER_DAY );
     }
     else
     {
-        t._msecs = (_msecs + ms) % MSECS_PER_DAY;
+        t._msecs = static_cast<unsigned>( (_msecs + ms) % MSECS_PER_DAY );
     }
 
     return t;
